Move important functions to `base_scraper.py``
